add strategy to bundle + check if kumulos is initialised in InAppTaskService + always register lifecycle watcher
