+++ kumulos/src/main/java/com/kumulos/android/AnalyticsContract.java
@@ -4,6 +4,7 @@ import android.app.Activity;
 import android.app.Application;
 import android.content.ContentValues;
 import android.content.Context;
+import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.database.sqlite.SQLiteDatabase;
@@ -339,16 +340,43 @@ import java.util.concurrent.atomic.AtomicBoolean;
         WeakReference<Context> mContextRef;
         static AtomicBoolean startNewSession;
 
+        //IN APP
+        private static WeakReference<Activity> currentActivityRef = new WeakReference<>(null);
+        @Nullable
+        static Activity getCurrentActivity() {
+            return currentActivityRef.get();
+        }
+        private static int numStarted = 0;
+        static boolean isBackground(){
+            return numStarted == 0;
+        }
+
+
         ForegroundStateWatcher(Context context) {
             mContextRef = new WeakReference<>(context);
             startNewSession = new AtomicBoolean(true);
         }
 
         @Override
-        public void onActivityCreated(Activity activity, Bundle savedInstanceState) { /* noop */ }
+        public void onActivityCreated(Activity activity, Bundle savedInstanceState) {
+            currentActivityRef = new WeakReference<Activity>(activity);
+        }
 
         @Override
-        public void onActivityStarted(Activity activity) { /* noop */ }
+        public void onActivityStarted(Activity activity) {
+            Integer tickleId = this.getTickleId(activity);
+            if ((isBackground() || tickleId != null) && KumulosInApp.isInAppEnabled()) {
+                InAppMessageService.readMessages(activity, isBackground(), tickleId);
+            }
+
+            numStarted++;
+        }
+
+        private Integer getTickleId(Activity activity){
+            Intent i = activity.getIntent();
+            int tickleIdExtra = i.getIntExtra(PushBroadcastReceiver.EXTRAS_KEY_TICKLE_ID, -1);
+            return tickleIdExtra == -1 ? null : tickleIdExtra;
+        }
 
         @Override
         public void onActivityResumed(Activity activity) {
@@ -401,13 +429,26 @@ import java.util.concurrent.atomic.AtomicBoolean;
         }
 
         @Override
-        public void onActivityStopped(Activity activity) { /* noop */ }
+        public void onActivityStopped(Activity activity) {
+            numStarted = Math.max(numStarted-1, 0);
+        }
 
         @Override
         public void onActivitySaveInstanceState(Activity activity, Bundle outState) {  /* noop */ }
 
         @Override
-        public void onActivityDestroyed(Activity activity) {  /* noop */ }
+        public void onActivityDestroyed(Activity activity) {
+            InAppMessagePresenter.maybeCloseDialog(activity);
+
+            Activity currentActivity = getCurrentActivity();
+            if (currentActivity == null){
+                return;
+            }
+
+            if (currentActivity.hashCode() == activity.hashCode()) {
+                currentActivityRef = new WeakReference<>(null);
+            }
+        }
     }
 
 }
