+++ kumulos/src/main/java/com/kumulos/android/InAppJavaScriptInterface.java
@@ -39,7 +39,7 @@ class InAppJavaScriptInterface {
 
         switch(messageType){
             case "READY":
-                Activity ca = InAppActivityLifecycleWatcher.getCurrentActivity();
+                Activity ca = AnalyticsContract.ForegroundStateWatcher.getCurrentActivity();
                 if (ca == null){
                     return;
                 }
@@ -47,7 +47,7 @@ class InAppJavaScriptInterface {
                 InAppMessagePresenter.clientReady(ca);
                 return;
             case "MESSAGE_OPENED":
-                Activity currentActivity = InAppActivityLifecycleWatcher.getCurrentActivity();
+                Activity currentActivity = AnalyticsContract.ForegroundStateWatcher.getCurrentActivity();
                 if (currentActivity == null){
                     return;
                 }
@@ -68,7 +68,7 @@ class InAppJavaScriptInterface {
 
     private void executeActions(List<ExecutableAction> actions){
 
-        Activity currentActivity = InAppActivityLifecycleWatcher.getCurrentActivity();
+        Activity currentActivity = AnalyticsContract.ForegroundStateWatcher.getCurrentActivity();
         if (currentActivity == null){
             return;
         }
