+++ kumulos/src/main/java/com/kumulos/android/InAppTaskService.java
@@ -2,6 +2,8 @@ package com.kumulos.android;
 
 import android.app.Application;
 import android.content.Context;
+import android.os.Bundle;
+
 import com.google.android.gms.gcm.GcmNetworkManager;
 import com.google.android.gms.gcm.GcmTaskService;
 import com.google.android.gms.gcm.PeriodicTask;
@@ -12,6 +14,7 @@ public class InAppTaskService extends GcmTaskService {
     private static final String TAG = "inapp-fetch";
     private static final long DESIRED_TASK_FREQUENCY = 15 * 60L;//15 minutes
     private static final long ACCEPTED_FREQUENCY_DEVIATION = 5 * 60L; //can run 5 min earlier
+    static final String KEY_CONFIG = "config";
 
     void startPeriodicFetches(Context context){
         long periodSecs = DESIRED_TASK_FREQUENCY;
@@ -21,12 +24,17 @@ public class InAppTaskService extends GcmTaskService {
             flexSecs = 15L;
         }
 
+        KumulosConfig config = Kumulos.getConfig();
+        Bundle bundle = new Bundle();
+        bundle.putBundle(KEY_CONFIG, config.toBundle());
+
         PeriodicTask periodic = new PeriodicTask.Builder()
                 .setService(InAppTaskService.class)
                 .setPeriod(periodSecs)
                 .setFlex(flexSecs)
                 .setTag(TAG)
                 .setUpdateCurrent(true)
+                .setExtras(bundle)
                 .build();
 
         new Thread(new Runnable() {
@@ -46,6 +54,14 @@ public class InAppTaskService extends GcmTaskService {
 
     @Override
     public int onRunTask(TaskParams params) {
+        if (!Kumulos.isInitialized()) {
+            Bundle bundle = params.getExtras();
+            Bundle configBundle = bundle.getBundle(KEY_CONFIG);
+
+            KumulosConfig config = KumulosConfig.fromBundle(configBundle);
+            Kumulos.initialize(this.getApplication(), config);
+        }
+
         boolean success = InAppMessageService.fetch(this);
         if (!success){
             return GcmNetworkManager.RESULT_RESCHEDULE;
